Crate libcrux_sha2

Modules

hacl

The generated hacl code This module contains generated hacl code.

Structs

Sha224

use it if we want to use hacl

Sha256

use it if we want to use hacl

Sha384

use it if we want to use hacl

Sha512

use it if we want to use hacl

Sha224Hash

A struct that implements libcrux_traits::digest traits.

Sha256Hash

A struct that implements libcrux_traits::digest traits.

Sha384Hash

A struct that implements libcrux_traits::digest traits.

Sha512Hash

A struct that implements libcrux_traits::digest traits.

Enums

Algorithm

use it if we want to use hacl The different Sha2 algorithms.

Constants

SHA224_LENGTH

The length of a SHA224 hash in bytes.

SHA256_LENGTH

The length of a SHA256 hash in bytes.

SHA384_LENGTH

The length of a SHA384 hash in bytes.

SHA512_LENGTH

The length of a SHA512 hash in bytes.

Traits

Digest

Re-export the Digest trait. A Hash algorithm returning hashes of length HASH_LEN.

Functions

sha224

use it if we want to use hacl SHA2 224 Will panic if payload is longer than u32::MAX to ensure that hacl-rs can process it.

sha256

use it if we want to use hacl SHA2 256 Will panic if payload is longer than u32::MAX to ensure that hacl-rs can process it.

sha384

use it if we want to use hacl SHA2 384 Will panic if payload is longer than u32::MAX to ensure that hacl-rs can process it.

sha512

use it if we want to use hacl SHA2 512 Will panic if payload is longer than u32::MAX to ensure that hacl-rs can process it.

Type Aliases

Sha224Hasher

A hasher for Sha224Hash.

Sha256Hasher

A hasher for Sha256Hash.

Sha384Hasher

A hasher for Sha384Hash.

Sha512Hasher

A hasher for Sha512Hash.