libcrux_sha2

Module hacl

The generated hacl code This module contains generated hacl code.

Constants

h224

h256

h384

h512

k224_256

k384_512

Functions

copy_256

Copies the state passed as argument into a newly allocated state (deep copy). The state is to be freed by calling free_256. Cloning the state this way is useful, for instance, if your control-flow diverges and you need to feed more (different) data into the hash in each branch.

copy_512

Copies the state passed as argument into a newly allocated state (deep copy). The state is to be freed by calling free_512. Cloning the state this way is useful, for instance, if your control-flow diverges and you need to feed more (different) data into the hash in each branch.

digest_224

Write the resulting hash into output, an array of 28 bytes. The state remains valid after a call to digest_224, meaning the user may feed more data into the hash via update_224.

digest_256

Write the resulting hash into output, an array of 32 bytes. The state remains valid after a call to digest_256, meaning the user may feed more data into the hash via update_256. (The digest_256 function operates on an internal copy of the state and therefore does not invalidate the client-held state p.)

digest_384

Write the resulting hash into output, an array of 48 bytes. The state remains valid after a call to digest_384, meaning the user may feed more data into the hash via update_384.

digest_512

Write the resulting hash into output, an array of 64 bytes. The state remains valid after a call to digest_512, meaning the user may feed more data into the hash via update_512. (The digest_512 function operates on an internal copy of the state and therefore does not invalidate the client-held state p.)

hash_224

Hash input, of len input_len, into output, an array of 28 bytes.

hash_256

Hash input, of len input_len, into output, an array of 32 bytes.

hash_384

Hash input, of len input_len, into output, an array of 48 bytes.

hash_512

Hash input, of len input_len, into output, an array of 64 bytes.

malloc_224

malloc_256

Allocate initial state for the SHA2_256 hash. The state is to be freed by calling free_256.

malloc_384

malloc_512

reset_224

reset_256

Reset an existing state to the initial hash state with empty data.

reset_384

reset_512

sha256_finish

sha256_init

sha256_update_last

sha256_update_nblocks

sha384_finish

sha384_init

sha384_update_last

sha384_update_nblocks

sha512_finish

sha512_init

sha512_update_last

sha512_update_nblocks

update_224

update_256

Feed an arbitrary amount of data into the hash. This function returns 0 for success, or 1 if the combined length of all of the data passed to update_256 (since the last call to reset_256) exceeds 2^61-1 bytes.

update_384

update_512

Feed an arbitrary amount of data into the hash. This function returns 0 for success, or 1 if the combined length of all of the data passed to update_512 (since the last call to reset_512) exceeds 2^125-1 bytes.

Type Aliases

state_t_224

state_t_256

state_t_384

state_t_512